magit-buffer-revision-hash: Move definition and make permanent local
The latter isn't necessary at this point but will be soon when we
start setting its value elsewhere.  In the long run it will become
unnecessary again once we remove some existing code that will become
obsolete in the near run.  Don't deal with the broken alignment for
now.  Fun times.