magit-mode-setup-internal: Call magit-xref-setup directly
We are going to stop using `magit-refresh-args', which means that
the buffer arguments have to be passed to `magit-xref-setup' as
function arguments.  That means it can no longer be called by
`magit-mode-hook', because that is a regular function.

Add a new variable `magit-xref-modes'.  Setting that to nil is
the new way of disabling this feature.  This can also be used by
third-party packages to add additional modes.